fix(review): establish one-way recipe section contracts
